+++ tkesdk/chktrans.go
@@ -7,11 +7,13 @@
 //
 // Date          Initials        Description
 // 05/03/2021    CLH             Initial version
+// 07/23/2021    CLH             Change message when a signature key cannot be used
 
 package tkesdk
 
 import (
 	"errors"
+	"os"
 	"strings"
 
 	"github.com/IBM/ibm-hpcs-tke-sdk/common"
@@ -83,6 +85,12 @@ func checkInputs(hc HsmConfig) ([]string, error) {
 	if len(hc.Admins) > 8 {
 		problems = append(problems, "No more than 8 administrators can be specified.")
 	}
+	if hc.FailoverUnits != 0 && (hc.FailoverUnits < 2 || hc.FailoverUnits > 3) {
+		problems = append(problems, "You must have either 2 or 3 Failover Units.")
+	}
+	// if hc.FailoverUnits > hc.units {
+	// 	problems = append(problems, "Cannot have more failover units than operational units")
+	// }
 
 	allKeysValid := true
 	for _, admin := range hc.Admins {
@@ -90,7 +98,16 @@ func checkInputs(hc HsmConfig) ([]string, error) {
 			problems = append(problems, "An administrator name is too long.  Names must be 30 characters or less.")
 		}
 		if !validKey(admin) {
-			problems = append(problems, "The signature key associated with "+admin.Name+" could not be accessed.")
+			ssURL := os.Getenv("TKE_SIGNSERV_URL")
+			if ssURL != "" {
+				problems = append(problems, "The signature key associated with "+
+					admin.Name+" could not be accessed.  An attempt was made "+
+					"to use a signing service.  The signing service may not be "+
+					"running at the specified URL and port.")
+			} else {
+				problems = append(problems, "The signature key associated with "+
+					admin.Name+" could not be accessed.")
+			}
 			allKeysValid = false
 		}
 	}
@@ -170,8 +187,8 @@ func internalCheckTransition(ci CommonInputs, hc HsmConfig,
 	for i := 0; i < len(hsminfo); i++ {
 
 		keepSKIs := make([]string, 0)
-		addSKIs  := make([]string, 0)
-		rmvSKIs  := make([]string, 0)
+		addSKIs := make([]string, 0)
+		rmvSKIs := make([]string, 0)
 
 		// What administrators do we want to keep or remove?
 		for j := 0; j < len(hsminfo[i].Admins); j++ {
@@ -197,8 +214,8 @@ func internalCheckTransition(ci CommonInputs, hc HsmConfig,
 		}
 
 		allKeepSKIs = append(allKeepSKIs, keepSKIs)
-		allAddSKIs  = append(allAddSKIs, addSKIs)
-		allRmvSKIs  = append(allRmvSKIs, rmvSKIs)
+		allAddSKIs = append(allAddSKIs, addSKIs)
+		allRmvSKIs = append(allRmvSKIs, rmvSKIs)
 
 		// Check whether the changes are possible
 		if len(keepSKIs) < hsminfo[i].SignatureThreshold {
@@ -296,6 +313,9 @@ func internalCheckTransition(ci CommonInputs, hc HsmConfig,
 			}
 		}
 	}
+
+	// TODO: need to check that we aren't removing failover hsms
+	// like if newFailover < currentFailover: problem
 
 	return problems, nil, allKeepSKIs, allAddSKIs, allRmvSKIs
 }