+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 <!-- Use FEATURES, ENHANCEMENTS, BUG FIXES as categories describing content. -->
 <!-- Put most recent change at top. -->
 
+## 1.0.2 (July 23, 2021)
+
+BUG FIXES:
+
+* Fix private endpoints, some message changes
+
+## 1.0.1 (July 16, 2021)
+
+FEATURES:
+
+* Add partial support for failover crypto units
+
 ## 1.0.0 (July 6, 2021)
 
 FEATURES: